Basic implementation
First public commit.

This is a rollup of about 2.5 weeks worth of tinkering, and contains
the player, parser, encoder, and two sample projects.

This was originally composed of many, many more commits, but it wasn't
until this point that I cleaned up and organized everything, added
license and copyright information, etc.

The emulator patches, hardware designs, and server details are still
works-in-progress.

There are still a couple of bugs in the player, and missing features
in the streaming interface.