Improve uC board assembly further
Change the register file CD74HC670M96 with SN74AHC373PWR and an
inverter, which is more easily available from JLCPCB.

This is not an exact replacement, as we replaced 2x 4x4 register files
with 2x 8-bit latches.  So instead of 4x 8-bit registers, we now only
have 2x 8-bit registers, but we only used 2 to begin with.

The new register hardware required us to change the address decoding
and the actual addresses of the registers.  Instead of 0xA13000 and
0xA13002, we now write to 0xA13010 and 0xA13012.