RLE compression
I initially experimented with LZ4 compression due to its speed.  In
Gangnam Style, for example, LZ4 reduced the file size by 49%, but
didn't noticeably improve overall throughput.

The LZ4 library needs memory, as much as 8kB (significant for a
microcontroller), and wants to write all output to a buffer in memory.
I believe it is the overhead of copying memory around that kills the
benefit of reducing the network traffic.

However, a simple RLE compression scheme reduced the file size by only
36%, but boosted throughput from 3.0Mbps to 5.1Mbps.  The RLE
decompression algorithm can operate on just a few bytes of extra
memory and can output directly to SRAM over serial, so it is more
efficient overall.