Fix SRAM CE signaling
It turns out that a PSRAM's ~{CE} signal may not stay on at all times.
There is a maximum pulse width of 15us for this chipset, so that the
DRAM internal to the PSRAM can go through a refresh cycle.

Tying ~{CE} to ~{WE} (for writes) or ~{SEGA_CE} (for reads) seems to
meet all the timing requirements.  This has yet to be verified,
though, for a lack of test pads for all ~{CE} signals on the previous
board revision.

(PCB routing in progress, will commit later.)