Fix color quantization and palette size in encoder
The encoder was not quantizing colors before palette generation, so palettes could contain similar colors that would not be representable as distinct on the Sega.  This has the effect of reducing our effective palette size.

We also had an off-by-one error in palette generation.  We asked ffmpeg to give us 15 colors in the palette, since the Sega palette has a 16th (actually 0th) color that must be fully transparent.  But ffmpeg had an on-by-default option to reserve a color for transparency, so we really only got 14 colors max... and because of the lack of quantization, it was possibly even fewer.

This should maximize color quality of our demo catalog after a re-encode, without needing any changes in the player, firmware, or hardware.

Thanks to the OP of joeyparrish/sega-slides#2 for pointing out a similar error in that project, leading to this discovery and fix.