Fix color quantization again
It turns out that only 3 bits of data per channel are actually used, even though the in-memory format is 4 bits per channel.  The low bit is ignored in each field, and only 512 colors are possible.

All demo content will now be re-encoded to fix quality issues.  The updates to the player folder here are in documentation only.

See also joeyparrish/sega-slides#2