Switch to a stable SGDK build
The latest from ghcr is broken and produces builds that can't read
controller input on a real Sega.  (They work fine on an emulator.)

The author of SGDK has deleted all workflows for his own docker
images, and now this is the recommended docker image to use.  As of
today, it works.  I'm pinning to a specific SHA256 to avoid any issues
with tags getting updated (as happened often with "latest" on the ghcr
image).