Fix video quantization
Quantizing the video should be done _again_ before palette
application.  Also, the quantization formula was simplified to reflect
how tile encoding will be done in a later step.  This greatly
increases quality and accuracy.

This was driven by comparisons between debug dumps and emulator
playback.  A screen cap and the "eye dropper" tool in GIMP was used to
analyze the colors in the debug video and the colors in the emulator,
to figure out how the formula went wrong.