Improve encoded audio quality
It turns out that dynamically normalizing audio really reduces quantization errors, especially on long-form content.  It also speeds up encoding.

In this change, we also start downmixing in a filter, rather than using -ac 1, so that we can normalize after all mixing.  This improves loudness.

Finally, we drop the noise filters and audio dithering, which were not actually helping much.